Allow users to set the aws region
For instances of the TDS relying on IAM profiles attached to EC2
instances in AWS, it's important for the TDS to know about the region
is operating within. This can be done through the use of a Java System
Property. This PR enables users to set an environment variable within
their compose.env file, the value of which gets set as a Java System
Property on startup of Tomcat.